+++ CHANGELOG.md
@@ -10,6 +10,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 - update to 2021 edition
 - update MSRV to 1.72.0
+
+
+- **breaking**: `dotenvy::Result` is now private
+- **breaking**: deprecate `dotenvy::var`, `dotenvy::from_filename*` 
+- `Error` is now `From<std::io::Error`
+- `errors` module renamed to `err`
+
 - **breaking**: dotenvy CLI uses `from_path` instead of `from_filename`
 - **breaking**: dotenvy CLI defaults to *./.env*, no longer traversing parent directories.
 - **breaking**: dotenvy CLI exits with code 2 instead of code 1 if the external command is omitted